Move the Ethers fetcher into its own directory
Also:

- Move the standard NFT fetching mechanism into its own file.
- Move cryptopunks.ts into the Ethers fetcher directory.
- Move some types from the Ethers fetcher into a separate file.
