+++ src/index.tsx
@@ -1,4 +1,4 @@
 export { NftProvider, useNft } from "./core"
 export { parseNftUrl } from "./utils"
-export { ethersFetcher } from "./fetchers/ethers-fetcher"
+export { ethersFetcher } from "./fetchers/ethers"
 export type { Loader, NftMetadata } from "./types"
